Bad solution for 2023-08-2
Given enough time, it would probably reach a solution.

It passes the tests.

I cheated, and looked at discussions of other people's solutions.
It seems there is an unwritten rule that each ghost will only pass
through one Z node in an instruction cycle and it will be at the point
of the final instruction. Therefore you can identify the cycle times
and work out when all the cycle times coincide.

I need to decide if I make my solution work for any reasonable input,
or if I encode in this assumption.

Web Assembly Text is not a good language for rapid prototyping and
inspection required to notice these patterns.